# packages/contracts/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Getter API",
    "version": "0.0.0",
    "description": "Generated from the TypeBox contracts in packages/contracts by packages/contracts/scripts/generate.ts. Do not edit by hand. JSON snake_case, ISO-8601 UTC timestamps, UUID identifiers. The API derives the workspace from the authenticated principal and never trusts a client workspace_id."
  },
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jg_session",
        "description": "HttpOnly session cookie issued by POST /auth/login."
      },
      "deviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Scoped device token for the extension or local runner; expires after 30 days."
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator worker credential for the internal task protocol only."
      }
    },
    "schemas": {
      "AcceptedResponse": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "properties": {
          "task_id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "const": "queued",
            "type": "string"
          }
        }
      },
...
      "CreateProfileImportRequest": {
        "additionalProperties": false,
        "type": "object",
        "properties": {
          "file_id": {
            "format": "uuid",
            "type": "string"
          },
          "pasted_text": {
            "minLength": 1,
            "maxLength": 200000,
            "type": "string"
          },
          "format_hint": {
            "anyOf": [
              {
                "const": "pdf",
                "type": "string"
              },
              {
                "const": "docx",
                "type": "string"
              },
              {
                "const": "linkedin_export_text",
                "type": "string"
              },
              {
                "const": "plain_text",
                "type": "string"
              },
              {
                "const": "auto",
                "type": "string"
              }
            ]
          }
        }
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "anyOf": [
                  {
                    "const": "VALIDATION_ERROR",
                    "type": "string"
                  },
                  {
                    "const": "MALFORMED_REQUEST",
                    "type": "string"
                  },
                  {
                    "const": "UNAUTHENTICATED",
                    "type": "string"
                  },
                  {
                    "const": "FORBIDDEN",
                    "type": "string"
                  },
                  {
                    "const": "NOT_FOUND",
                    "type": "string"
                  },
                  {
                    "const": "CONFLICT",
                    "type": "string"
                  },
                  {
                    "const": "STALE_REVISION",
                    "type": "string"
                  },
                  {
                    "const": "IDEMPOTENCY_MISMATCH",
                    "type": "string"
                  },
                  {
                    "const": "PAYLOAD_TOO_LARGE",
                    "type": "string"
                  },
                  {
                    "const": "UNPROCESSABLE",
                    "type": "string"
                  },
                  {
                    "const": "QUOTA_EXCEEDED",
                    "type": "string"
                  },
                  {
                    "const": "SETUP_CLOSED",
                    "type": "string"
                  },
                  {
                    "const": "BUDGET_EXHAUSTED",
                    "type": "string"
                  },
                  {
                    "const": "PROVIDER_UNAVAILABLE",
                    "type": "string"
                  },
                  {
                    "const": "INTERNAL_ERROR",
                    "type": "string"
                  }
                ]
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "patternProperties": {
                  "^(.*)$": {
                    "type": "string"
                  }
                }
              },
              "request_id": {
                "format": "uuid",
                "type": "string"
              }
            }
          }
        }
      },
...
      "Profile": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "revision",
          "confirmed_revision",
          "contact",
          "locale",
          "facts",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "revision": {
            "minimum": 1,
            "type": "integer"
          },
          "confirmed_revision": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContactValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "locale": {
            "anyOf": [
              {
                "const": "en",
                "type": "string"
              },
              {
                "const": "es",
                "type": "string"
              }
            ]
          },
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileFact"
            }
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/profile/imports": {
      "post": {
        "operationId": "createProfileImport",
        "summary": "Queue extraction from an uploaded document or pasted text.",
        "tags": [
          "profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "202": {
            "description": "Queue extraction from an uploaded document or pasted text.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Mandatory for this command. Replaying the same key returns the stored response for 24 hours; reusing it with a different body returns 409.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileImportRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/profile/imports/{id}": {
      "get": {
        "operationId": "getProfileImport",
        "summary": "Extraction draft, warnings, conflicts and evidence.",
        "tags": [
          "profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Extraction draft, warnings, conflicts and evidence.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileImportView"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/profile/imports/{id}/confirm": {
      "post": {
        "operationId": "confirmProfileImport",
        "summary": "Promote user-accepted drafts into confirmed facts.",
        "tags": [
          "profile"
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Promote user-accepted drafts into confirmed facts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request body, query or header.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated: no valid session, device token or worker credential.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted for this workspace resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Absent or not accessible from the authenticated workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict: stale revision, duplicate command or idempotency key reuse with a different body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Payload exceeds the configured upload limit.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Structurally valid but invalid domain input.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Quota or rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "description": "Anti-CSRF token for cookie-authenticated mutations; origin is verified too.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 200
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImportRequest"
              }
            }
          }
        }
      }
    },
...
  }
}